backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.38.0
pydantic==2.9.2
